refactor: Move <AssociatedInstitution> to the shared "/components" directory for easier reuse
