--- src/pages/Filing/ViewUserProfile/AssociatedInstitutions.less
@@ -1,14 +1,0 @@
-.institution-list {
-  .cf-icon-svg {
-    margin-right: 5px;
-  }
-
-  .status-label {
-    margin-right: 10px;
-    font-weight: 500;
-  }
-}
-
-.green {
-  color: #20aa3f;
-}
